// src/api/__generated__/mining/model/totalRewardDTO.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Mining API
 * 挖矿服务接口
 * OpenAPI spec version: v1.0
 */

/**
 * 累计奖励响应
 */
export interface TotalRewardDTO {
  /**
   * 累计派遣挖矿奖励
   * @minimum 0
   */
  totalMiningReward?: number;
  /**
   * 累计邀请奖励
   * @minimum 0
   */
  totalInviteReward?: number;
}