// src/api/__generated__/mining/model/weeklyStatsDTO.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Mining API
 * 挖矿服务接口
 * OpenAPI spec version: v1.0
 */

/**
 * 周统计数据响应
 */
export interface WeeklyStatsDTO {
  /**
   * 本周总奖池
   * @minimum 0
   */
  weekPool?: number;
  /**
   * 本周邀请总奖池
   * @minimum 0
   */
  weekInvitePool?: number;
  /**
   * 用户本周名义产出
   * @minimum 0
   */
  weeklyNominalOutput?: number;
  /**
   * 本周预估实际产出
   * @minimum 0
   */
  weeklyTotalOutput?: number;
}