// src/api/__generated__/wallet/model/inviteInfoResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama User Wallet API
 * 用户、钱包、任务中心接口
 * OpenAPI spec version: v1.0
 */

/**
 * 邀请信息
 */
export interface InviteInfoResponse {
  /** 当前用户ID */
  userId?: string;
  /** 当前用户邀请码 */
  inviteCode?: string;
  /** 邀请人用户ID，没有则为空 */
  inviterUserId?: string;
  /** 邀请绑定时间，没有则为空 */
  bindAt?: string;
  /** 总邀请人数 */
  totalInviteCount?: string;
  /** 有效用户数（待激活数 = 总邀请人数 − 有效用户数） */
  validUserCount?: string;
}